% Generated by roxygen2 (4.0.1): do not edit by hand
\name{capitalise}
\alias{capitalise}
\alias{capitalize}
\title{Capitalise string}
\usage{
capitalise(x)

capitalize(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\description{
Turn the first letter of a string into its capital version
}
\examples{
capitalise("the quick red fox jumps over the lazy brown dog")
}
\seealso{
\code{\link[base]{toupper}}, \code{\link{capitalise_words}}
}

