% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{case}
\alias{case}
\alias{str_to_lower}
\alias{str_to_title}
\alias{str_to_upper}
\title{Convert case of a string.}
\usage{
str_to_upper(string)

str_to_lower(string)

str_to_title(string)
}
\arguments{
\item{string}{String to modify}
}
\description{
Convert case of a string.
}
\examples{
dog <- "The quick brown dog"
str_to_upper(dog)
str_to_lower(dog)
str_to_title(dog)
}

