\name{check_pattern}
\alias{check_pattern}
\title{Check that pattern is of the correct type for stringr functions...}
\usage{check_pattern(pattern, string, replacement)}

\description{
  Check that pattern is of the correct type for stringr functions
}
\keyword{internal}
\arguments{
  \item{pattern}{input vector}
}
