% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interp.R
\name{eval_interp_matches}
\alias{eval_interp_matches}
\title{Evaluate String Interpolation Matches}
\usage{
eval_interp_matches(matches, env)
}
\arguments{
\item{matches}{Match data}

\item{env}{The environment in which to evaluate the expressions.}
}
\value{
A character vector of replacement strings.
}
\description{
The expression part of string interpolation matches are evaluated in a
specified environment and formatted for replacement in the original string.
Used internally by \code{\link{str_interp}}.
}
\author{
Stefan Milton Bache
}

