% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interp.R
\name{extract_expressions}
\alias{extract_expressions}
\title{Extract Expression Objects from String Interpolation Matches}
\usage{
extract_expressions(matches)
}
\arguments{
\item{matches}{Match data}
}
\value{
list of R expressions
}
\description{
An interpolation match object will contain both its wrapping \code{${ }} part
and possibly a format. This extracts the expression parts and parses them to
prepare them for evaluation.
}
\author{
Stefan Milton Bache
}

