% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interp.R
\name{extract_formats}
\alias{extract_formats}
\title{Extract String Interpolation Formats from Matched Placeholders}
\usage{
extract_formats(matches)
}
\arguments{
\item{matches}{Match data}
}
\value{
A character vector of format specifiers.
}
\description{
An expression placeholder for string interpolation may optionally contain a
format valid for \code{\link{sprintf}}. This function will extract such or
default to "s" the format for strings.
}
\author{
Stefan Milton Bache
}

