% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interp.R
\name{interp_placeholders}
\alias{interp_placeholders}
\title{Match String Interpolation Placeholders}
\usage{
interp_placeholders(string)
}
\arguments{
\item{string}{character: The string to be interpolated.}
}
\value{
list containing \code{indices} (regex match data) and \code{matches},
  the string representations of matched expressions.
}
\description{
Given a character string a set of expression placeholders are matched. They
are of the form \code{${...}} or optionally \code{$[f]{...}} where \code{f}
is a valid format for \code{\link{sprintf}}.
}
\author{
Stefan Milton Bache
}

