% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.r
\name{like_converter}
\alias{like_converter}
\title{Pattern Conversion Between LIKE and Regex structures}
\usage{
like_converter(pattern)
}
\arguments{
\item{pattern}{SQL LIKE operator pattern to be converted.}
}
\value{
A converted pattern.
}
\description{
Pattern Conversion Between LIKE and Regex structures
}
\examples{
like_converter("bana\%")
like_converter("app_e")
}
