% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.r
\name{like_to_regex}
\alias{like_to_regex}
\title{Pattern Conversion Between LIKE and Regex structures}
\usage{
like_to_regex(pattern)
}
\arguments{
\item{pattern}{SQL LIKE operator pattern to be converted.}
}
\value{
A converted pattern.
}
\description{
Turns a SQL LIKE pattern into a recognisible regular expression pattern.
}
