% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interp.R
\name{match_brace}
\alias{match_brace}
\title{Utility Function for Matching a Closing Brace}
\usage{
match_brace(opening, closing)
}
\arguments{
\item{opening}{integer: Vector with positions of opening braces.}

\item{closing}{integer: Vector with positions of closing braces.}
}
\value{
Integer with the posision of the matching brace.
}
\description{
Given positions of opening and closing braces \code{match_brace} identifies
the closing brace matching the first opening brace.
}
\author{
Stefan Milton Bache
}

