% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/modifiers.r
\name{modifier-deprecated}
\alias{ignore.case}
\alias{modifier-deprecated}
\alias{perl}
\title{Deprecated modifier functions.}
\usage{
ignore.case(string)

perl(pattern)
}
\description{
Please use \code{\link{regexp}} and \code{\link{coll}} instead.
}
\keyword{internal}

