% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conv.R
\name{str_conv}
\alias{str_conv}
\title{Specify the encoding of a string.}
\usage{
str_conv(string, encoding)
}
\arguments{
\item{string}{String to re-encode.}

\item{encoding}{Name of encoding. See \code{\link[stringi]{stri_enc_list}}
for a complete list.}
}
\description{
This is a convenient way to override the current encoding of a string.
}
\examples{
# Example from encoding?stringi::stringi
x <- rawToChar(as.raw(177))
x
str_conv(x, "ISO-8859-2") # Polish "a with ogonek"
str_conv(x, "ISO-8859-1") # Plus-minus
}

