% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{str_dedent}
\alias{str_dedent}
\title{str_dedent}
\usage{
str_dedent(text)
}
\arguments{
\item{text}{\code{character} The input string or character vector.}
}
\value{
The input string or character vector with leading indentation removed.
}
\description{
Remove common leading indentation from strings
}
\note{
This function is similar to Python's \code{dedent} function in the \code{textwrap} library. It removes common leading indentation from strings.
}
\examples{
dedent("  Hello\n    World")
# Expected Output: "Hello\n  World"

dedent("  Line 1\n  Line 2\n  Line 3")
# Expected Output: "Line 1\nLine 2\nLine 3"

dedent("No indentation")
# Expected Output: "No indentation"

dedent(
    "
    this
    is
        a
    test
    "
)
# Expected Output: "\nthis\nis\n    a\ntest\n"
}
\author{
chrimaho
}
