% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.r
\name{str_dice}
\alias{str_dice}
\alias{str_dice<-}
\title{Dice a character vector into pieces of fixed length.}
\usage{
str_dice(string, width = 1L)

str_dice(string, width = 1L) <- value
}
\arguments{
\item{string}{input character vector.}

\item{width}{integer giving the width in letters of each piece to return.}
}
\value{
A character vector of substrings each of length \code{width}, except
the last will be shorter if \code{length(string)} is not a multiple of \code{width}.
}
\description{
Wraps stri_sub for situations where completely decomposing a string
into fixed-length substrings is appropriate. The last
}
\details{
This is useful for example for breaking a coding sequence of DNA
into codon-sized triplets.
}
\examples{
hw <- "Hadley Wickham"
CDS <- "ATGTTTGGGTAG"

str_dice(hw, width=1L)
str_dice(hw, width=2L)
str_dice(hw, width=4L)
str_dice(CDS, width=3L)
}
\seealso{
The underlying implementation in \code{\link[stringi:stri_sub]{stringi::stri_sub()}}
}
