% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.r
\name{str_explode}
\alias{str_explode}
\title{Explode a string into single characters}
\usage{
str_explode(string)
}
\arguments{
\item{string}{Input string as a list or character vector (or something
coercible to one)}
}
\description{
Returns the same class as the input \code{string}: either a \code{list} or an atomic
vector.
}
\examples{
str_explode("apple")
str_explode(c("apple", "banana"))
str_explode(list("apple", "banana"))
}
