\name{str_extract}
\alias{str_extract}
\title{Extract pieces of a string that match a pattern.}
\usage{str_extract(string, pattern)}

\description{
  Extract pieces of a string that match a pattern.
}
\value{list of character vectors.}
\arguments{
  \item{string}{input character vector}
  \item{pattern}{pattern to look for.  See \code{\link{regex}} for
description.}
}
