% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{str_flatten}
\alias{str_flatten}
\title{Flatten a string}
\usage{
str_flatten(string, collapse = "", last = NULL, na.rm = FALSE)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{collapse}{String to insert between each piece. Defaults to \code{""}.}

\item{last}{Optional string use in place of final separator.}

\item{na.rm}{Remove missing values? If \code{FALSE}, the result will be \code{NA}
if any element of \code{string} is \code{NA}.}
}
\value{
A string, i.e. a character vector of length 1
}
\description{
Reduce a character vector to a single string. This is a summary function
because regardless of the length of the input \code{x}, it always returns a
single string.
}
\examples{
str_flatten(letters)
str_flatten(letters, "-")

str_flatten(letters[1:4], ", ", ", and ")
str_flatten(letters[1:3], ", ", ", and ")
str_flatten(letters[1:2], ", ", ", and ")
str_flatten(letters[1], ", ", ", and ")
str_flatten(letters[0], ", ", ", and ")
}
