% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/py_format.R
\name{str_format}
\alias{str_format}
\title{format a string like python}
\usage{
str_format(str, ...)
}
\arguments{
\item{str}{A string}

\item{...}{See details}
}
\value{
A string or a string vector
}
\description{
Format a string
}
\examples{
#Numeric placeholders
str_format("Hello {1}, Hello {2}", "world", "data")
str_format("Hello {1}, Hello {2}", c("world", "data"))
str_format("Hello {1}, Hello {2}", list("world", "data"))

#Named placeholders
str_format("Hello {name}, Hello {name_}", name="world", name_="data")
str_format("Hello {name}, Hello {name_}", list(name="world", name_="data"))
str_format("Hello {name}, Hello {name_}", c(name="world", name_="data"))

#vector paramters
str_format("Hello {name}, Hello {name_}", name=paste0("World", 1:10), name_=paste0("Data", 1:10))
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.format}
\url{https://github.com/nicolewhite/pystr/blob/master/R/pystr_format.R}
}

