% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{str_ilike}
\alias{str_ilike}
\title{Detect a pattern in the same way as \code{SQL}'s \code{ILIKE} operator}
\usage{
str_ilike(string, pattern)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{A character vector containing a SQL "like" pattern.
See above for details.}
}
\value{
A logical vector the same length as \code{string}.
}
\description{
\code{str_ilike()} follows the conventions of the SQL \code{ILIKE} operator:
\itemize{
\item Must match the entire string.
\item \verb{_} matches a single character (like \code{.}).
\item \verb{\%} matches any number of characters (like \verb{.*}).
\item \verb{\\\%} and \verb{\\_} match literal \verb{\%} and \verb{_}.
\item The match is case insensitive.
}
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
str_ilike(fruit, "app")
str_ilike(fruit, "app\%")
str_ilike(fruit, "APP\%")
str_ilike(fruit, "ba_ana")
str_ilike(fruit, "\%apple")
}
