% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{str_indent}
\alias{str_indent}
\title{Indent multiple strings}
\usage{
str_indent(x, indent_level = 3, indent_character = " ", collapse = NULL)
}
\arguments{
\item{indent_level}{controls the number of characters to add,}

\item{indent_character}{controls what character is placed.}
}
\value{
a newline separated list of words with a 'tab' character in between
}
\description{
\code{str_indent()} adds a customizable amount of spaces (or other characters)
before every word, as if it was tabbed, and puts every word on a new line.
}
\examples{
x <- c("why", "video", "cross", "extra", "deal", "authority")
# Default options
str_indent(x)
# Extra indentation
str_indent(x, indent_level = 4)
# Using periods for visibility
str_indent(x, indent_character = '.')
# Both parameters changed
str_indent(x, indent_level = 6, indent_character = '-')
}
\seealso{
\code{\link[=str_spaceout]{str_spaceout()}} to add spaces between words.
}
