% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/left.R
\name{str_left}
\alias{str_left}
\alias{str_right}
\title{Left- or right-most element(s)}
\usage{
str_left(x, n = 1)

str_right(x, n = 1)
}
\arguments{
\item{x}{a vector that you wish to extract from.}

\item{n}{an optional integer indicating the number of
elements you wish to extract from \code{x} starting from the
left-most (\code{left}) or right-most (\code{right}) element.}
}
\value{
A character vector containing the \code{n} left-most
(\code{str_left}) or right-most (\code{str_right}) elements
of \code{x}.
}
\description{
Extract the \code{n} left-most (\code{str_left}) or right-most
(\code{str_right}) elements of \code{x}.
}
\examples{
x <- c("abcde", "12345")
  str_left(x)
  str_left(x, n = 3)
  str_right(x)
  str_right(x, n = 2)
}
\author{
Gerhard Nachtmann \email{kpm.nachtmann@gmail.com}
}
\seealso{
\code{\link{stri_sub}}, \code{\link{str_sub}},
         \code{\link{substr}}
}
\keyword{left}
\keyword{right}

