% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_length}
\alias{str_length}
\title{The length of a string.}
\usage{
str_length(string)
}
\arguments{
\item{string}{input vector. This must be an atomic vector, and will be
coerced to a character vector}
}
\value{
numeric vector giving number of characters in each element of the
  character vector. Missing string have missing length.
}
\description{
Technically this returns the number of "code points", in a string. One
code point usually corresponds to one character, but not always. For example,
an u with a umlaut might be represent as a single character or as the
combination a u and an umlaut.
}
\examples{
str_length(letters)
str_length(NA)
str_length(factor("abc"))
str_length(c("i", "like", "programming", NA))
}
\seealso{
\code{\link[stringi]{stri_length}} which this function wraps.
}
\keyword{character}

