% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{str_match}
\alias{str_match}
\title{Extract first matched group from a string.}
\usage{
str_match(string, pattern)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for, as defined by an ICU regular
expression. See \link[stringi]{stringi-search-regex} for more details.}
}
\value{
Character matrix. First column is the complete match, followed by
  one column for each capture group.
}
\description{
Vectorised over \code{string} and \code{pattern}.
}
\examples{
strings <- c(" 219 733 8965", "329-293-8753 ", "banana", "595 794 7569",
  "387 287 6718", "apple", "233.398.9187  ", "482 952 3315",
  "239 923 8115", "842 566 4692", "Work: 579-499-7527", "$1000",
  "Home: 543.355.3679")
phone <- "([2-9][0-9]{2})[- .]([0-9]{3})[- .]([0-9]{4})"

str_extract(strings, phone)
str_match(strings, phone)
}
\seealso{
\code{\link{str_extract}} to extract the complete match,
  \code{\link{str_match_all}} to extract all matches,
  \code{\link[stringi]{stri_match_first_regex}} for the underlying
  implementation.
}
\keyword{character}

