% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{str_match_all}
\alias{str_match_all}
\title{Extract all matched groups from a string.}
\usage{
str_match_all(string, pattern)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for.

  The default interpretation is a regular expression, as described
  in \link[stringi]{stringi-search-regex}. Control options with
  \code{\link{regex}()}.

  Match a fixed string (i.e. by comparing only bytes), using
  \code{\link{fixed}(x)}. This is fast, but approximate. Generally,
  for matching human text, you'll want \code{\link{coll}(x)} which
  respects character matching rules for the specified locale.

  Match character, word, line and sentence boundaries with
  \code{\link{boundary}()}. An empty pattern, "", is equivalent to
  \code{boundary("character")}.}
}
\value{
A list of character matrices. First column is the complete match,
  followed by one column for each capture group.
}
\description{
Vectorised over \code{string} and \code{pattern}.
}
\examples{
strings <- c("Home: 219 733 8965.  Work: 229-293-8753 ",
  "banana pear apple", "595 794 7569 / 387 287 6718")
phone <- "([2-9][0-9]{2})[- .]([0-9]{3})[- .]([0-9]{4})"

str_extract_all(strings, phone)
str_match_all(strings, phone)
}
\seealso{
\code{\link{str_extract_all}} to extract the complete match,
  \code{\link{str_match}} to extract just the first match,
  \code{\link[stringi]{stri_match_all_regex}} for the underlying
  implementation.
}

