% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{str_order}
\alias{str_order}
\alias{str_sort}
\title{Order or sort a character vector}
\usage{
str_order(
  x,
  decreasing = FALSE,
  na_last = TRUE,
  locale = stringr_locale(),
  numeric = FALSE,
  ...
)

str_sort(
  x,
  decreasing = FALSE,
  na_last = TRUE,
  locale = stringr_locale(),
  numeric = FALSE,
  ...
)
}
\arguments{
\item{x}{A character vector to sort.}

\item{decreasing}{A boolean. If \code{FALSE}, the default, sorts from
lowest to highest; if \code{TRUE} sorts from highest to lowest.}

\item{na_last}{Where should \code{NA} go? \code{TRUE} at the end,
\code{FALSE} at the beginning, \code{NA} dropped.}

\item{locale}{Locale to use for comparisons. See
\code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}} for all possible options.
Defaults to \code{"en"} (American English) to ensure that the default collation
is consistent across platforms. See \code{\link[=stringr_locale]{stringr_locale()}} for information on
how to control this default.}

\item{numeric}{If \code{TRUE}, will sort digits numerically, instead
of as strings.}

\item{...}{Other options used to control collation. Passed on to
\code{\link[stringi:stri_opts_collator]{stringi::stri_opts_collator()}}.}
}
\description{
\code{str_order()} returns an integer vector that returns the desired order
when used for subsetting; \code{str_sort()} returns the sorted vector.
}
\examples{
str_order(letters)
str_sort(letters)

str_order(letters, locale = "haw")
str_sort(letters, locale = "haw")

x <- c("100a10", "100a5", "2b", "2a")
str_sort(x)
str_sort(x, numeric = TRUE)
}
\seealso{
\code{\link[stringi:stri_order]{stringi::stri_order()}} for the underlying implementation.
}
