% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{str_pad}
\alias{str_pad}
\title{Pad a string.}
\usage{
str_pad(string, width, side = c("left", "right", "both"), pad = " ")
}
\arguments{
\item{string}{input character vector}

\item{width}{pad strings to this minimum width}

\item{side}{side on which padding character is added (left, right or both)}

\item{pad}{single padding character (default is a space)}
}
\value{
character vector
}
\description{
Vectorised over \code{string}, \code{width} and \code{pad}.
}
\examples{
rbind(
  str_pad("hadley", 30, "left"),
  str_pad("hadley", 30, "right"),
  str_pad("hadley", 30, "both")
)

# All arguments are vectorised except side
str_pad(c("a", "abc", "abcdef"), 10)
str_pad("a", c(5, 10, 20))
str_pad("a", 10, pad = c("-", "_", " "))

# Longer strings are returned unchanged
str_pad("hadley", 3)
}
\seealso{
\code{\link{str_trim}} to remove whitespace
}
\keyword{character}

