% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{str_replace_na}
\alias{str_replace_na}
\title{Turn NA into "NA"}
\usage{
str_replace_na(string, replacement = "NA")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{replacement}{A single string.}
}
\description{
Turn NA into "NA"
}
\examples{
str_replace_na(c(NA, "abc", "def"))
}
