\name{str_reverse}
\alias{str_reverse}
\title{Returns a string in reverse order.}
\usage{
  str_reverse(string)
}
\arguments{
  \item{string}{input character vector}
}
\value{
  character vector.
}
\description{
  The characters in each elements of a vector are reversed,
  e.g. "gnat" -> "tang"
}
\examples{
str_reverse("tang")
}
\seealso{
  \code{\link{str_sort}} to sort a string
}
\keyword{character}

