% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{str_space}
\alias{str_space}
\title{Split up strings}
\usage{
str_space(x, split_style = "default", split_char = " ")
}
\arguments{
\item{split_style}{controls the placement of splits and can be
modified to the following styles:
\code{acronyms}: splits at lowercase-uppercase and between two uppercase letters
followed by lowercase,
\code{allcaps}: splits between all consecutive uppercase letters
\code{default}: splits only when a lowercase letter is succeeded by an uppercase letter
\code{everything}: splits between every character}

\item{split_char}{controls the character that splits words.
This defaults to space but can be changed to anything.}
}
\description{
\code{str_space()} adds a space wherever a lowercase letter is followed by an
uppercase letter, with modifiable functionality for special cases.
}
\examples{
input <- "XMLHTTPRequestAndHTMLParser"
# Default splitting
str_space(input, split_style = "default")
# Split acronyms
str_space(input, split_style = "acronyms")
# Split all consecutive uppercase letters
str_space(input, split_style = "allcaps")
# Split every character
str_space(input, split_style = "everything")

}
\seealso{
\code{\link[=str_indent]{str_indent()}} to add tabs between words.
}
