% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{str_standardize}
\alias{str_standardize}
\title{Standardize}
\usage{
str_standardize(string, prefix = "var_")
}
\arguments{
\item{string}{a string}

\item{prefix}{a string}
}
\value{
a standardized string
}
\description{
Standardize removes accents, blank spaces, special caracters from a string
}
\details{
See http://stackoverflow.com/a/36898175/1967500 for the tricks to remove accents
}
\examples{
str_standardize("2017/07/07", prefix = "date_")
str_standardize("code externe de l'action")

}
