% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.r
\name{str_starts}
\alias{str_starts}
\title{Detect the presence or absence of a pattern at the beginning of a string.}
\usage{
str_starts(string, pattern, negate = FALSE)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern with which the string starts. More details at
\code{\link{str_detect}}.}

\item{negate}{If \code{TRUE}, return non-matching elements.}
}
\value{
A logical vector.
}
\description{
Vectorised over \code{string} and \code{pattern}.
Equivalent to \code{grepl(paste0("^", pattern), x)}.
}
\examples{
fruit <- c("apple", "banana", "pear", "pinapple")
str_starts(fruit, "p")
str_starts(fruit, "p", negate = TRUE)
}
\seealso{
\code{\link[=str_detect]{str_detect()}} which this function wraps.
}
