% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{str_to_pascal}
\alias{str_to_pascal}
\alias{str_to_camel}
\alias{str_to_snake}
\alias{str_to_kebab}
\title{Convert between different types of programming case}
\usage{
str_to_pascal(string, locale = "en")

str_to_camel(string, locale = "en")

str_to_snake(string, separator = "_", locale = "en")

str_to_kebab(string, locale = "en")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{locale}{Locale to use for comparisons. See
\code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}} for all possible options.
Defaults to "en" (English) to ensure that default behaviour is
consistent across platforms.}

\item{separator}{string. For snake case, can change the \verb{_}}
}
\description{
\itemize{
\item \code{str_to_pascal()} converts to pascal case, where only the first letter of
each word is capitalized, with no separation between
words.
\item \code{str_to_camel()} converts to camel case, where only the first letter of
each word after the first word is capitalized, with no separation between
words.
\item \code{str_to_kebab()} converts to kebab case, where words are converted to
lower case and separated by dashes (\code{-}).
\item \code{str_to_snake()} converts to snake case, where words are converted to
lower case and separated by underscores (\verb{_}).
}
}
