% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad-trim.r
\name{str_trim_all}
\alias{str_trim_all}
\title{Trim whitespace from start, middle, and end of string.}
\usage{
str_trim_all(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector.
}
\description{
Trim whitespace from start, middle, and end of string.
}
\examples{
str_trim_all("  String with trailing,  middle, and leading white space\\t")
str_trim_all("\\n\\nString with excess,  trailing and leading white   space\\n\\n")
}
\seealso{
\code{\link{str_trim}} to see end-only trimming
}
