% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{str_unique}
\alias{str_unique}
\title{Keep unique strings only.}
\usage{
str_unique(string, ...)
}
\arguments{
\item{string}{A character vector to return unique entries.}

\item{...}{Other options used to control matching behavior between duplicate
strings. Passed on to \code{stringi::stri_opts_collator()}.}
}
\value{
A character vector.
}
\description{
Vectorized over \code{string}.
}
\examples{
str_unique(letters[c(1:3, 2, 3)])

str_unique(c("motley", "mötley", "pinguino", "pingüino"))
str_unique(c("motley", "mötley", "pinguino", "pingüino"), strength = 1)
}
\seealso{
\code{\link[=unique]{unique()}}, \code{\link[stringi:stri_unique]{stringi::stri_unique()}} which this function wraps.
}
