% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{str_view}
\alias{str_view}
\alias{str_view_all}
\title{View HTML rendering of regular expression match}
\usage{
str_view(
  string,
  pattern = NULL,
  match = NA,
  html = getOption("stringr.html", TRUE)
)

str_view_all(
  string,
  pattern = NULL,
  match = NA,
  html = getOption("stringr.html", TRUE)
)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described
in \link[stringi:about_search_regex]{stringi::about_search_regex}. Control options with
\code{\link[=regex]{regex()}}.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[=fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[=coll]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[=boundary]{boundary()}}. An empty pattern, "", is equivalent to
\code{boundary("character")}.}

\item{match}{If \code{TRUE}, shows only strings that match the pattern.
If \code{FALSE}, shows only the strings that don't match the pattern.
Otherwise (the default, \code{NA}) displays both matches and non-matches.}

\item{html}{Use HTML output? If \code{TRUE} will create an HTML widget;
if \code{FALSE} will style using ANSI escapes.}
}
\description{
\code{str_view()} shows the first match; \code{str_view_all()} shows all
the matches. To build regular expressions interactively, check out the
\href{https://www.garrickadenbuie.com/project/regexplain/}{RegExplain RStudio addin}.
}
\examples{
# Show special characters
str_view(c("\"\\\\", "\\\\\\\\\\\\", "fgh"))

# Show first match
str_view(c("abc", "def", "fgh"), "[aeiou]")
str_view(c("abc", "def", "fgh"), "^")
str_view(c("abc", "def", "fgh"), "..")

# Show all matches
str_view_all(c("abc", "def", "fgh"), "d|e")
}
