% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stringr-data}
\alias{fruit}
\alias{sentences}
\alias{stringr-data}
\alias{words}
\title{Sample character vectors for practicing string manipulations.}
\format{A character vector.}
\usage{
sentences

fruit

words
}
\description{
\code{fruit} and \code{word} come from the \code{rcorpora} package
written by Gabor Csardi; the data was collected by Darius Kazemi
and made available at \url{https://github.com/dariusk/corpora}.
\code{sentences} is a collection of "Harvard sentences" used for
standardised testing of voice.
}
\keyword{datasets}

