% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{stringr_locale}
\alias{stringr_locale}
\title{Locale used by stringr}
\usage{
stringr_locale()
}
\description{
\code{stringr_locale()} returns a single string representing the default locale
used by stringr in functions like \code{\link[=str_sort]{str_sort()}} and \code{\link[=str_to_upper]{str_to_upper()}}. The
default locale is \code{"en"} (American English) on all platforms, but can be
customized by setting the global option, \code{tidyverse.locale_collation}, to a
locale found in \code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}}.

We only recommend setting the global option if you don't expect your code
to be shared across multiple computers. Otherwise, it is more reproducible
to set the \code{locale} argument directly.
}
\examples{
stringr_locale()

withr::with_options(
  list(tidyverse.locale_collation = "fr"),
  stringr_locale()
)
}
