% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/case.R
\name{case}
\alias{case}
\alias{str_to_lower}
\alias{str_to_title}
\alias{str_to_upper}
\title{Convert case of a string.}
\usage{
str_to_upper(string, locale = "")

str_to_lower(string, locale = "")

str_to_title(string, locale = "")
}
\arguments{
\item{string}{String to modify}

\item{locale}{Locale to use for translations.}
}
\description{
Convert case of a string.
}
\examples{
dog <- "The quick brown dog"
str_to_upper(dog)
str_to_lower(dog)
str_to_title(dog)

# Locale matters!
str_to_upper("i", "en") # English
str_to_upper("i", "tr") # Turkish
}

