% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modifiers}
\alias{coll}
\alias{fixed}
\alias{modifiers}
\alias{regex}
\title{Control matching behaviour with modifier functions.}
\usage{
fixed(pattern)

coll(pattern, ignore_case = FALSE, locale = NULL, ...)

regex(pattern, ignore_case = FALSE, multiline = FALSE, comments = FALSE,
  dotall = FALSE, ...)
}
\arguments{
\item{string}{string to match exactly as is}
}
\description{
Control matching behaviour with modifier functions.
}
\examples{
pattern <- "a.b"
strings <- c("abb", "a.b")
str_detect(strings, pattern)
str_detect(strings, fixed(pattern))
}
\keyword{character}

