% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{str_dedent}
\alias{str_dedent}
\title{Remove common leading indentation from strings}
\usage{
str_dedent(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}
}
\value{
A character vector the same length as \code{string}
}
\description{
\code{str_dedent()} is designed to make it possible to correctly indent multiline
strings inside of function calls, while generating the desired amount of
whitespace in the output.

It does this by removing the common leading indentation from each line
(ignoring lines only containing whitespace), and removing the first line,
if it only contains whitespace.

It is inspired by Python's
\href{https://docs.python.org/3/library/textwrap.html#textwrap.dedent}{\code{textwrap.dedent()}}.
}
\examples{
str_dedent("
   Hello
   World
")

f <- function() {
  str_dedent("
    Line 1
    Line 2
    Line 3
  ")
}
cat(str_dedent(f()))
}
