% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{str_dedent}
\alias{str_dedent}
\title{Remove common leading indentation from strings}
\usage{
str_dedent(text)
}
\arguments{
\item{text}{\code{character} The input string or character vector.}
}
\value{
The input string or character vector with leading indentation removed.
}
\description{
This function is similar to Python's \code{dedent} function in the \code{textwrap}
library. It removes common leading indentation from strings.
}
\examples{
str_dedent("  Hello\n    World")

str_dedent("  Line 1\n  Line 2\n  Line 3")

str_dedent("No indentation")

str_dedent(
    "
    this
    is
        a
    test
    "
)
}
