\name{str_dup}
\alias{str_dup}
\title{Duplicate strings within a character vector.}
\usage{str_dup(string, times)}

\description{
  Duplicate strings within a character vector.
}
\value{character vector}
\keyword{internal}
\arguments{
  \item{string}{input character vector}
  \item{times}{number of times to duplicate each string}
}
