% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dup.r
\name{str_dup}
\alias{str_dup}
\title{Duplicate and concatenate strings within a character vector.}
\usage{
str_dup(string, times)
}
\arguments{
\item{string}{Input character vector.}

\item{times}{Number of times to duplicate each string.}
}
\value{
A character vector.
}
\description{
Vectorised over \code{string} and \code{times}.
}
\examples{
fruit <- c("apple", "pear", "banana")
str_dup(fruit, 2)
str_dup(fruit, 1:3)
str_c("ba", str_dup("na", 0:5))
}

