% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{str_locate}
\alias{str_locate}
\title{Locate the position of the first occurence of a pattern in a string.}
\usage{
str_locate(string, pattern)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for.

  The default interpretation is a regular expression, as described
  in \link[stringi]{stringi-search-regex}. Control options with
  \code{\link{regex}()}.

  Match a fixed string (i.e. by comparing only bytes), using
  \code{\link{fixed}(x)}. This is fast, but approximate. Generally,
  for matching human text, you'll want \code{\link{coll}(x)} which
  respects character matching rules for the specified locale.

  Match character, word, line and sentence boundaries with
  \code{\link{boundary}()}. An empty pattern, "", is equivalent to
  \code{boundary("character")}.}
}
\value{
Integer matrix. First column gives start postion of match, and
  second column gives end position.
}
\description{
Vectorised over \code{string} and \code{pattern}.
}
\examples{
fruit <- c("apple", "banana", "pear", "pinapple")
str_locate(fruit, "$")
str_locate(fruit, "a")
str_locate(fruit, "e")
str_locate(fruit, c("a", "b", "p", "p"))
}
\seealso{
\code{\link[stringi]{stri_locate_first}} which this function wraps,
  \code{\link{str_extract}} for a convenient way of extracting matches,
  \code{\link{str_locate_all}} to locate position of all matches
}

