\name{str_match}
\alias{str_match}
\title{Extract first matched group from a string.}
\usage{str_match(string, pattern)}

\description{
  Extract first matched group from a string.
}
\value{character matrix. First column is the complete match, followed by 
one for each capture group}
\keyword{character}
\arguments{
  \item{string}{input character vector}
  \item{pattern}{with groups}
}
