\name{str_match_all}
\alias{str_match_all}
\title{Extract all matched groups from a string...}
\usage{str_match_all(string, pattern)}

\description{
  Extract all matched groups from a string
}
\value{list of character matrices, as given by \code{\link{str_match}}}
\keyword{character}
\arguments{
  \item{string}{input character vector}
  \item{pattern}{with groups}
}
