% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sort.R
\name{str_order}
\alias{str_order}
\alias{str_sort}
\title{Order or sort a character vector.}
\usage{
str_order(x, decreasing = FALSE, na_last = TRUE, locale = "", ...)

str_sort(x, decreasing = FALSE, na_last = TRUE, locale = "", ...)
}
\arguments{
\item{x}{A character vector to sort.}

\item{decreasing}{A boolean. If \code{FALSE}, the default, sorts from
lowest to highest; if \code{TRUE} sorts from highest to lowest.}

\item{na_last}{Where should \code{NA} go? \code{TRUE} at the end,
\code{FALSE} at the beginning, \code{NA} dropped.}

\item{locale}{In which locale should the sorting occur? Defaults to
the current locale.}

\item{...}{Other options used to control sorting order. Passed on to
\code{\link[stringi]{stri_opts_collator}}.}
}
\description{
Order or sort a character vector.
}
\examples{
str_order(letters, locale = "en")
str_sort(letters, locale = "en")

str_order(letters, locale = "haw")
str_sort(letters, locale = "haw")
}
\seealso{
\code{\link[stringi]{stri_order}} for the underlying implementation.
}

