% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{str_replace}
\alias{str_replace}
\title{Replace first occurrence of a matched pattern in a string.}
\usage{
str_replace(string, pattern, replacement)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for.

  The default interpretation is a regular expression, as described
  in \link[stringi]{stringi-search-regex}. Control options with
  \code{\link{regex}()}.

  Match a fixed string (i.e. by comparing only bytes), using
  \code{\link{fixed}(x)}. This is fast, but approximate. Generally,
  for matching human text, you'll want \code{\link{coll}(x)} which
  respects character matching rules for the specified locale.}

\item{replacement}{replacement string.  References of the form \code{\1},
\code{\2} will be replaced with the contents of the respective matched
group (created by \code{()}) within the pattern.}
}
\value{
character vector.
}
\description{
Vectorised over \code{string}, \code{pattern} and \code{replacement}.
}
\examples{
fruits <- c("one apple", "two pears", "three bananas")
str_replace(fruits, "[aeiou]", "-")
str_replace_all(fruits, "[aeiou]", "-")

str_replace(fruits, "([aeiou])", "")
str_replace(fruits, "([aeiou])", "\\\\1\\\\1")
str_replace(fruits, "[aeiou]", c("1", "2", "3"))
str_replace(fruits, c("a", "e", "i"), "-")
}
\seealso{
\code{\link{sub}} which this function wraps,
  \code{\link{str_replace_all}} to replace all matches
}
\keyword{character}

