% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_replace_all}
\alias{str_replace_all}
\title{Replace all occurrences of a matched pattern in a string.}
\usage{
str_replace_all(string, pattern, replacement)
}
\arguments{
\item{string}{input vector. This must be an atomic vector, and will be
coerced to a character vector}

\item{pattern,replacement}{Supply separate pattern and replacement strings
  to vectorise over the patterns. References of the form \code{\1},
  \code{\2} will be replaced with the contents of the respective matched
  group (created by \code{()}) within the pattern.

  If you want to apply multiple patterns and replacements to each string,
  pass a named character to \code{pattern}.}
}
\value{
character vector.
}
\description{
Vectorised over \code{string}, \code{pattern} and \code{replacement}.
Shorter arguments will be expanded to length of longest.
}
\examples{
fruits <- c("one apple", "two pears", "three bananas")
str_replace(fruits, "[aeiou]", "-")
str_replace_all(fruits, "[aeiou]", "-")

str_replace_all(fruits, "([aeiou])", "")
str_replace_all(fruits, "([aeiou])", "\\\\1\\\\1")
str_replace_all(fruits, "[aeiou]", c("1", "2", "3"))
str_replace_all(fruits, c("a", "e", "i"), "-")

# If you want to apply multiple patterns and replacements to the same
# string, pass a named version to pattern.
str_replace_all(str_c(fruits, collapse = "---"),
 c("one" = 1, "two" = 2, "three" = 3))
}
\seealso{
\code{\link{gsub}} which this function wraps,
  \code{\link{str_replace}} to replace a single match
}
\keyword{character}

