% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace.r
\name{str_replace_na}
\alias{str_replace_na}
\title{Turn NA into "NA"}
\usage{
str_replace_na(string, replacement = "NA")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{replacement}{Supply separate pattern and replacement strings
  to vectorise over the patterns. References of the form \code{\1},
  \code{\2} will be replaced with the contents of the respective matched
  group (created by \code{()}) within the pattern.

  For \code{str_replace_all} only, you can perform multiple patterns and
  replacements to each string, by passing a named character to
  \code{pattern}.}
}
\description{
Turn NA into "NA"
}
\examples{
str_replace_na(c("NA", "abc", "def"))
}

