% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pad-trim.r
\name{str_trim}
\alias{str_trim}
\title{Trim whitespace from start and end of string.}
\usage{
str_trim(string, side = c("both", "left", "right"))
}
\arguments{
\item{string}{A character vector.}

\item{side}{Side on which to remove whitespace (left, right or both).}
}
\value{
A character vector.
}
\description{
Trim whitespace from start and end of string.
}
\examples{
str_trim("  String with trailing and leading white space\\t")
str_trim("\\n\\nString with trailing and leading white space\\n\\n")
}
\seealso{
\code{\link{str_pad}} to add whitespace
}

