% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{str_unique}
\alias{str_unique}
\title{Remove duplicated strings}
\usage{
str_unique(string, ...)
}
\arguments{
\item{string}{A character vector to return unique entries.}

\item{...}{Other options used to control matching behavior between duplicate
strings. Passed on to \code{\link[stringi:stri_opts_collator]{stringi::stri_opts_collator()}}.}
}
\value{
A character vector.
}
\description{
\code{str_unique()} removes duplicated values, with optional control over
how duplication is measured.
}
\examples{
str_unique(c("a", "b", "c", "b", "a"))

# Use ... to pass additional arguments to stri_unique()
str_unique(c("motley", "mötley", "pinguino", "pingüino"))
str_unique(c("motley", "mötley", "pinguino", "pingüino"), strength = 1)
}
\seealso{
\code{\link[=unique]{unique()}}, \code{\link[stringi:stri_unique]{stringi::stri_unique()}} which this function wraps.
}
