% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/view.R
\name{str_view}
\alias{str_view}
\title{View HTML rendering of regular expression match.}
\usage{
str_view(x, pattern)
}
\arguments{
\item{pattern}{Pattern to look for.

  The default interpretation is a regular expression, as described
  in \link[stringi]{stringi-search-regex}. Control options with
  \code{\link{regex}()}.

  Match a fixed string (i.e. by comparing only bytes), using
  \code{\link{fixed}(x)}. This is fast, but approximate. Generally,
  for matching human text, you'll want \code{\link{coll}(x)} which
  respects character matching rules for the specified locale.

  Match character, word, line and sentence boundaries with
  \code{\link{boundary}()}. An empty pattern, "", is equivalent to
  \code{boundary("character")}.}
}
\description{
View HTML rendering of regular expression match.
}
\examples{
str_view(c("abc", "def", "fgh"), "[aeiou]")
str_view(c("abc", "def", "fgh"), "^")
}

