% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr.R
\docType{package}
\name{stringr-package}
\alias{stringr}
\alias{stringr-package}
\title{stringr: Simple, Consistent Wrappers for Common String Operations}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A consistent, simple and easy to use set of
    wrappers around the fantastic 'stringi' package. All function and
    argument names (and positions) are consistent, all functions deal with
    "NA"'s and zero length vectors in the same way, and the output from
    one function is easy to feed into the input of another.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://stringr.tidyverse.org}
  \item \url{https://github.com/tidyverse/stringr}
  \item Report bugs at \url{https://github.com/tidyverse/stringr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com} [copyright holder]

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
